requirements.txt
pytest==7.4.4
pytest-html==4.1.1
requests==2.31.0
jsonschema==4.21.1
python-dotenv==1.0.0
email-validator
